requirements.txt
# langchain==0.2.16
# streamlit==1.38.0
# chromadb==1.0.15
# sentence-transformers==3.0.1
# google-generativeai==0.7.2
# python-dotenv==1.0.1


# Core Framework
streamlit==1.38.0
langchain==0.2.16

# Vector Database & Embeddings
chromadb==1.0.15
sentence-transformers==3.0.1

# AI/ML APIs
google-generativeai==0.7.2
openai>=1.0.0

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
python-pptx==0.6.23

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Utilities
python-dotenv==1.0.1
requests>=2.31.0
beautifulsoup4>=4.12.0

# Monitoring & Logging
psutil>=5.9.0
structlog>=23.1.0

# Security
cryptography>=41.0.0

# Performance
aiohttp>=3.8.0
asyncio-mqtt>=0.13.0

# Development (optional)
black>=23.0.0
flake8>=6.0.0
pytest>=7.4.0
pre-commit>=3.3.0
